Don't narrow float to long before asInteger
asInteger does more than just cast to long and will produce a
Bignum if necessary. This likely changed in CRuby when Bignum and
Fixnum merged into Integer.

+++ core/src/main/java/org/jruby/util/Sprintf.java
@@ -760,7 +760,7 @@ public class Sprintf {
                     case INTEGER: // no-op
                         break;
                     case FLOAT:
-                        arg = asInteger(context, ((RubyFloat) arg).asLong(context));
+                        arg = asInteger(context, ((RubyFloat) arg).asDouble(context));
                         break;
                     case STRING:
                         arg = ((RubyString) arg).stringToInum(0, true);
